Fix mongodb error handling in resource handlers
Single-field validation errors caused server 500's, because error.errors is None in that case.

Fixed that, factored MongoDB error handling into a decorator, and handle some other MongoDB errors.

This change adds the decorators but doesn't remove the previous `try` from `label_resources.py`, to avoid a merge conflict with a pending PR.

It adds error tests of varying comprehensiveness for labels and events. It doesn't test the changed code for ics and subscription resources, but the decorators are tested by the other test cases, and the error conditions for these other resources weren't tested before (so this isn't a coverage regression).