Refactor the splash screen CSS
This seemed like a good idea, but I'm not sure if it makes it clearer or less clear.

This renders the same as before, but with a couple(?) pixels less spacing next to the arrows. That's easily fixable, but I wasn't sure how intentional the exact pixel spacing was in the original.